Initial checkin
We just discovered the "All One" island. We set up a primitive shelter,
which will do for the first days. An old barrel we discovered at the
beach is the only piece of furniture so far. What else will we discover
behind the forest? What lies to the east? What shall we do with our
garden?